+++ decidim-initiatives/app/commands/decidim/initiatives/admin/update_initiative.rb
@@ -54,8 +54,13 @@ module Decidim
 
           if current_user.admin?
             add_admin_accessible_attrs(attrs)
-          elsif initiative.custom_signature_end_date_enabled? && initiative.created?
-            attrs[:signature_end_date] = form.signature_end_date
+          else
+            if initiative.custom_signature_end_date_enabled? && initiative.created?
+              attrs[:signature_end_date] = form.signature_end_date
+            end
+            if initiative.area_enabled? && initiative.created?
+              attrs[:decidim_area_id] = form.area_id
+            end
           end
 
           attrs
@@ -66,6 +71,7 @@ module Decidim
           attrs[:signature_end_date] = form.signature_end_date
           attrs[:offline_votes] = form.offline_votes if form.offline_votes
           attrs[:state] = form.state if form.state
+          attrs[:decidim_area_id] = form.area_id
 
           if initiative.published?
             @notify_extended = true if form.signature_end_date != initiative.signature_end_date &&