+++ decidim-initiatives/app/forms/decidim/initiatives/initiative_form.rb
@@ -12,6 +12,7 @@ module Decidim
       attribute :description, String
       attribute :type_id, Integer
       attribute :scope_id, Integer
+      attribute :area_id, Integer
       attribute :decidim_user_group_id, Integer
       attribute :signature_type, String
       attribute :signature_end_date, Date
@@ -21,6 +22,7 @@ module Decidim
       validates :title, length: { maximum: 150 }
       validates :signature_type, presence: true
       validates :type_id, presence: true
+      validates :area, presence: true, if: ->(form) { form.area_id.present? }
       validate :scope_exists
       validates :signature_end_date, date: { after: Date.current }, if: lambda { |form|
         form.context.initiative_type.custom_signature_end_date_enabled? && form.signature_end_date.present?
@@ -37,6 +39,10 @@ module Decidim
 
       def scope_id
         super.presence
+      end
+
+      def area
+        @area ||= current_organization.areas.find_by(id: area_id)
       end
 
       private