+++ decidim-initiatives/spec/system/admin/initiative_types_controller_spec.rb
@@ -52,7 +52,7 @@ describe "InitiativeTypesController", type: :system do
   context "when updating an initiative type" do
     let(:initiatives_type) do
       create(:initiatives_type, :online_signature_enabled, :attachments_disabled, :undo_online_signatures_enabled,
-             :custom_signature_end_date_disabled, organization: organization)
+             :custom_signature_end_date_disabled, :area_disabled, organization: organization)
     end
 
     it "Updates the initiative type" do
@@ -68,6 +68,7 @@ describe "InitiativeTypesController", type: :system do
       check "Enable attachments"
       uncheck "Enable participants to undo their online signatures"
       check "Enable authors to choose the end of signature collection period"
+      check "Enable authors to choose the area for their initiative"
 
       click_button "Update"